refactor transactions
