+++ DEVELOPER.md
@@ -69,6 +69,7 @@ Some folders to highlight:
 
 - `examine/recipe/` - this folder contains components for the _recipe area_ of the Examine page. The recipe area is the place where all the conflicts, conditions, trademarks, and history can be browsed. It's also where the user can select conflicts for their decision.
 - `examine/request_info/` - this folder contains components for the _request info_ section of the Examine page, which are the 5 cells near the top of the Examine page that contain information about the NR currently being examined.
+- `history/` - this folder contains base components for the transaction history and notification history page.
 
 If a component is generic enough that it can be used anywhere in the application (e.g. a button), then it can go directly under the `components` folder.
